libhsmd: Migrate handle_get_output_scriptpubkey
